Modified to set file size

+++ backend/migration/20230204112529_initial.up.sql
@@ -33,7 +33,7 @@ create table if not exists job_files
     is_output    boolean      not null default false,
     file_key     varchar(255) not null,
     file_name    varchar(255) not null,
-    size         integer      not null default 0,
+    size         bigint       not null default 0,
     content_type varchar(100) not null default 'application/octet-stream',
     created_at   timestamp    not null,
     updated_at   timestamp    not null,
